requirements.txt
# Receipt OCR App - Python Backend Requirements

# Core Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Image Processing
pillow>=9.0.0
opencv-python>=4.5.0

# OCR Services
google-cloud-vision>=3.4.0
pytesseract>=0.3.0

# Data Processing
pandas>=1.3.0
python-dateutil>=2.8.0

# File Handling
python-magic>=0.4.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.0.0

# Export Functionality
openpyxl>=3.1.0

# HTTP Client
requests>=2.31.0

# Testing Dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
httpx>=0.24.0
pytest-cov>=4.0.0